Restructuring of code.
