+++ jsons/deserializers/default_dict.py
@@ -1,5 +1,5 @@
 from typing import Optional, Callable
-from jsons._main_impl import load
+from jsons._load_impl import load
 
 
 def default_dict_deserializer(
