+++ jsons/deserializers/default_mapping.py
@@ -1,6 +1,7 @@
 from collections.abc import Mapping
 from typing import Mapping as MappingType
-from jsons._common_impl import get_naked_class
+
+from jsons._compatibility_impl import get_naked_class
 from jsons.deserializers import default_dict_deserializer
