+++ jsons/deserializers/default_union.py
@@ -1,7 +1,7 @@
 from typing import Union
 from jsons._common_impl import get_class_name
 from jsons._compatibility_impl import get_union_params
-from jsons._main_impl import load
+from jsons._load_impl import load
 from jsons.exceptions import JsonsError, DeserializationError
